+++ playwright.config.ts
@@ -33,13 +33,13 @@ export default defineConfig({
 
   use: {
     // Base URL to use in actions like `await page.goto('/')`.
-    baseURL: "http://127.0.0.1:3000",
+    baseURL: "http://127.0.0.1:5173",
 
     // Collect trace when on failed test.
     trace: "retain-on-failure",
     ignoreHTTPSErrors: true,
     // An object containing additional HTTP headers to be sent with every request.
-    // bypassCSP: true,
+    bypassCSP: true,
     launchOptions: {
       args: ["--disable-web-security", "--ignore-certificate-errors"],
     },
@@ -55,8 +55,8 @@ export default defineConfig({
   ],
   // Run your local dev server before starting the tests.
   webServer: {
-    command: "REACT_APP_STAGE=testing npm run start",
-    url: "http://127.0.0.1:3000",
+    command: "VITE_STAGE=testing npm run start",
+    url: "http://127.0.0.1:5173",
     reuseExistingServer: !process.env.CI,
   },
   expect: {